+++ tsconfig.json
@@ -15,20 +15,20 @@
     "jsx": "preserve",
     "incremental": true,
     "types": ["vite/client"],
-    "baseUrl": ".",
+    "baseUrl": "./src",
     "paths": {
-      "@/api/*": ["./src/api/*"],
-      "@/assets/*": ["./src/assets/*"],
-      "@/theme": ["./src/shared/theme"],
-      "@/components": ["./src/shared/components"],
-      "@/modules/*": ["./src/modules/*"],
-      "@/models/*": ["./src/models/*"],
-      "@/store": ["./src/store"],
-      "@/ui": ["./src/shared/ui"],
-      "@/utils": ["./src/shared/utils"],
-      "@/constants/*": ["./src/shared/constants/*"],
-      "@/mocks/*": ["./src/mocks/*"],
-      "@/hooks": ["./src/shared/hooks"]
+      "@/api/*": ["./api/*"],
+      "@/assets/*": ["./assets/*"],
+      "@/theme": ["./shared/theme"],
+      "@/components": ["./shared/components"],
+      "@/modules/*": ["./modules/*"],
+      "@/models/*": ["./models/*"],
+      "@/store": ["./store"],
+      "@/ui": ["./shared/ui"],
+      "@/utils": ["./shared/utils"],
+      "@/constants/*": ["./shared/constants/*"],
+      "@/mocks/*": ["./mocks/*"],
+      "@/hooks": ["./shared/hooks"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "lint-staged.config.js"],
